requirements.txt
pandas
mysql-connector-python
Flask
Werkzeug